requirements.txt
Babel==2.7.0
config==0.3.9
confusable-homoglyphs==3.2.0
dateparser==0.7.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.23
django-admin-shortcuts==2.0.0
django-allauth==0.40.0
django-bootstrap3==11.1.0
django-heroku==0.3.1
django-jet==1.0.8
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-registration==2.4.1
djangorestframework==3.10.3
gcloud==0.17.0
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.14.0
jet-bridge-base==0.6.6
jet-django==0.8.9
jws==0.1.3
oauth2client==3.0.0
oauthlib==3.1.0
phonenumbers==8.10.23
Pillow==6.2.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-decouple==3.1
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2019.3
regex==2019.11.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
tzlocal==2.0.0
wcwidth==0.1.7
whitenoise==4.0
